pyproject.toml
[project]
name = "sentinel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.2",
    "langchain>=1.2.13",
    "langchain-core>=1.2.23",
    "langchain-google-genai>=4.2.1",
    "langsmith>=0.7.22",
    "python-dotenv>=1.2.2",
    "supabase>=2.28.3",
    "torch>=2.11.0",
    "transformers>=5.4.0",
    "uvicorn>=0.42.0",
    "web3>=7.14.1",
]
